backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
asyncpg==0.29.0
pydantic==2.6.1
pydantic-settings==2.1.0
weaviate-client==4.4.1
celery==5.3.6
redis==5.0.1
requests==2.31.0
python-multipart==0.0.9
